Add macOS support (might not work) and update application version
This commit introduces several key changes to enhance macOS support for the application. A new job was added to the `publish.yml` file for publishing the macOS project and uploading the release asset. The `AlbionDataAvalonia.Desktop.MacOS.csproj` file was updated to change the output type and include properties for macOS app bundle creation.

The application version was updated from `0.12.1.0` to `0.12.2.0` in both the `inno.iss` and `AlbionDataAvalonia.csproj` files.

Additionally, improvements were made to the token refresh logic in `AuthService.cs` and `AFMUploader.cs`, and a new `Info.plist` file was added to define macOS application properties. A `run.sh` script was also created to set environment variables for executing the macOS application.